Allow specifying expected encoder len
This introduces new user methods to CategoricalEncoder: `expect_len` and `ignore_len`.
When neither of these are called, a UserWarning will be emitted once, informing the user that it is recommended to explicitly provide the category count they expect.

Solves #1589 and #1696 as `CTCTextEncoder` is a subclass of `CategoricalEncoder` (through `TextEncoder`).