Fix `parallel_map` bugs and refactor it
Most importantly, this fixes #1958 which was caused by the queue sometimes depleting fast enough. The fix is implemented by ensuring that the source iterator is actually depleted before we stop mapping.

This refactors a bunch of the `parallel_map` logic.
The function was now turned into an util class (though the intended way is still to invoke it through `parallel_map`).
This is partly because the linter was starting to complain about the complexity, and because the nested functions were starting to get confusing.
Objects from that class can still only be used once, but this clears up some of the logic.

Comments were added to clarify on the logic further.

Exceptions are now handled more gracefully. They are raised back in the scheduler much earlier when possible, and other futures in the pool will be cancelled should an exception be raised (this was already the case when not specifying an executor, as it would shut down the pool).