Pin torch and k2 version specifically in readthedocs-requirements.txt
We would rather not pin PyTorch and k2 in the docs-requirements (esp as the package to install depends on whether you want CPU or CUDA), so this is added to the new rtd requirements file instead. This is not ideal but I don't see much of a clean way around this.