`data/`->`meta/` + rename + remove test_manifest.txt from repo
Still need a further commit to download the test_manifest from dropbox once
I can reupload it.
The dataset download script was removed and merged into the README as we
always avoid having shell scripts in the repo (might be worth simplifying
the steps down the line and avoid the destructive renames etc.).
The extract_audio script was renamed to explicitly refer to the DSTC11 version.