Add simple wrappers for unet/diffwave for diffusion (#2598)
* Add simple wrappers for unet/diffwave for diffusion

Linter-induced refactors have had the consequences that some models which used `**kwargs` and discarded those would fail due to additional keyword arguments they did not expect.

The prior solution for this issue was #2523 but after some discussion with @tparcollet we agreed to prefer this wrapper approach.
The motivation towards this wrapper approach IMO is that the proposed solution to fix this at diffusion level would indeed, theoretically, work out of the box. However, to my understanding, it essentially relied on implicitly dropping extraneous parameters, even if they were passed through by the caller of the diffusion module. This is problematic, because this may not always result in an error in case of a typo (e.g. `length` vs `lengths`, where the wrapped module only takes it optionally). So we preferred a more explicit solution that is less likely to accidentally eat arguments.

* Fix `length` param name for audiomnist diffusion