Fix `download_file` gzip semantics
This change makes it so that `.gz` files that are **not** tarballs are uncompressed through `gzip`, and the rest gets extracted with `shutil.unpack_archive` under the assumption they are proper archive files.

The original comment states that `unpack_archive` does not work with `.tar.gz`, but this appears untrue. This may have originated from extraction of a `.txt.gz` file failing, due to it not being an archive. It seems that this comment induced another incorrect change that ended up breaking the semantics of `.gz` extraction again.