Try forcing omegaconf to 2.0.6 to see if doc building passes
The issue is that the pinned omegaconf version for latest releases of fairseq has dependency metadata that pip now consider invalids. This results in failed dependency solving when installing fairseq, because all candidate versions get rejected.
Hopefully, pinning a specific version of `omegaconf` as a direct dependency should force pip to consider this version instead of ignoring it for the above reason.