Rework doc landing page
License stuff and references is fine and all, but it occupies a lot of space. There is now much fewer text, which makes it a lot more obvious where to find tutorials, and outlines the structureof documentation quite better.

Also added sphinx_design for the license/reference dropdowns, which looks pretty fine