Replace `torchaudio.load` with `self.load_audio` when trivial
`load_audio` is preferred as it goes through common path handling code and normalization (for resampling and downmixing).

This relies on the audio normalizer being correctly configured for the model sample rate. Commit ccd0ed introduced functionality to infer the sample rate from `hparams.sample_rate` by default when the audio normalizer is not specified.

This should result in strictly identical behavior save for misconfigurations of the normalizer in hparams (if this code actually uses the proper tensor format that is).