URL fixes and URL check fixes (#2692)
* Fix URLs to `unstable-v0.6`, use relative link for `recipes/`

* Relative link in readme for PERFORMANCE.md

* URL to Spectral Clutsering in README dead, point to web archive copy

* Replace dead URL to `wham_noise.zip`

This AWS bucket link is used in `recipes/ESC50` and is linked to in http://wham.whisper.ai/ which seems to be treated as the official source for that dataset elsewhere in SB.

Not going to bother moving it to `fetch`.

* check_url.yaml rework, regex support, parallel, expanded scope

* Tutorial URL fix

* Fix links to inference code that was changed in 1.0

* Web archive BPE_Gage.pdf

* Remove broken link to PyTorch doc in tutorial

I am not sure what this is actually intended to point to, as the URL seemingly referred to `torchvision`. Removing for now...

* Fix link to doc in quaternion tutorial

* Fix link to papers in tutorial

* Fix Colab/GitHub URL for asr-metrics.ipynb

* Fix more tutorial dead links to the web archive

* Update ESC-50 dataset link and ignore dead URL false positive

* Ignore dead URL false positive in DNS

* Be more verbose about URL check errors

* Ignore URL check true positive for urbansounddataset

This is the actual README, i'll avoid tampering with it beyond adding a notice. Better indicate to the user the link might be dead.

* Fix format string typo

* Fix formatting

* Add the web archive to ignored URLs for URL checks

* Add arXiv to ignored URLs for URL checks

* Disable TLS verification in URL checks

This fixes the URL check for `https://sail.usc.edu`. We don't really care about the MITM risk here as we do nothing with the data.

* Add kaggle to URL exclusion regex

* VoxLingua107 pre-compiled shards are dead, add warning + ignore check

* Undo broken ')' handling for URL check, just ignore one URL for now

* Fix URL and typo in speech-classification-from-scratch

* Formatting

* Ignore TLS verify=False warning in URL check